Implemento de RestConroller para exponer API REST
Se exponen los servicios CRUD para la tabla de productos.
Ajuste en el tipo de dato para el ID de categoría y productos.
Correcciones en nombre de atributos de clases de dominio.

+++ src/main/java/com/platzi/market/domain/repository/ProductRepository.java
@@ -9,14 +9,14 @@ public interface ProductRepository {
 
     List<Product> getAll();
 
-    Optional<List<Product>> getByCategory(int categoryId);
+    Optional<List<Product>> getByCategory(long categoryId);
 
     Optional<List<Product>> getScarceProducts(int quantity);
 
-    Optional<Product> getProduct(int productId);
+    Optional<Product> getProduct(long productId);
 
     Product save(Product product);
 
-    void delete(int productId);
+    void delete(long productId);
 
 }
