+++ src/main/java/com/platzi/market/domain/service/ProductService.java
@@ -18,11 +18,11 @@ public class ProductService {
         return productRepository.getAll();
     }
 
-    public Optional<Product> getProduct(int productId) {
+    public Optional<Product> getProduct(long productId) {
         return productRepository.getProduct(productId);
     }
 
-    public Optional<List<Product>> getByCategory(int categoryId) {
+    public Optional<List<Product>> getByCategory(long categoryId) {
      return productRepository.getByCategory(categoryId);
     }
 
@@ -30,7 +30,7 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public boolean delete(int productId) {
+    public boolean delete(long productId) {
         return getProduct(productId).map(product -> {
             productRepository.delete(productId);
             return true;
