+++ src/main/java/com/platzi/market/persistence/crud/ProductoCrudRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductoCrudRepository extends CrudRepository<Producto, Integer> {
+public interface ProductoCrudRepository extends CrudRepository<Producto, Long> {
 
-    List<Producto> findByIdCategoriaOrderByNombreAsc(int idCategoria);
+    List<Producto> findByIdCategoriaOrderByNombreAsc(long idCategoria);
 
     Optional<List<Producto>> findByCantidadStockLessThanAndEstado(int cantidadStock, boolean estado);
